visualizations/requirements.txt
psycopg2-binary==2.9.10
pandas==2.2.3
matplotlib==3.9.3
seaborn==0.13.2
python-dotenv==1.0.1
